+++ CHANGELOG.md
@@ -5,7 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.0.7] - Unreleased
+## [0.1.0] - 2021-11-21
+
+- Give the owner a chance to tap device during update
+- Bump to version 0.1 so we can distinguish patch and breaking release in the future
+
+## [0.0.7] - 2021-11-21
 
 - Fix the Windows 10 bug (via `lpc55-host` bump)
 - Fix the incorrect udev rules file